fix: code review round 18 — compile errors, logic bugs, and script improvements
- Fix spawn_detached to require Send bound for tokio::spawn (native)
- Fix player.rs calling .read() on String instead of clone
- Fix bible/chapter.rs using wasm-only spawn_local (use dioxus::prelude::spawn)
- Fix mint_mgmt.rs and receive.rs u64->u32 type mismatch
- Fix download.rs missing return for fallback case
- Fix voice component prop name mismatch (on_recording_complete)
- Add User-Agent header to shared http_client()
- Add async IIFE wrapper for mobile clipboard Promise
- Add generation guard to blossom upload progress clear
- Add error logging for nip65 relay persistence
- Convert ics.rs download_ics to return Result instead of panicking
- Fix calendar date arithmetic (use f64 for JS Date set_date)
- Fix zap_dialog WebLN success check (verify boolean is true)
- Fix quotes.rs tests to use chrono_now_secs for native compatibility
- Add URL encoding for wavlake LNURL content_id and app_id
- Fix date_helpers days_in_month to return 0 for invalid month
- Improve build-android.sh: use find for NDK detection, single-quote trap
- Use shared http_client() in url_metadata.rs
- Add compile_error guards for wallet_database feature combos
- Add comment explaining mem::forget in pinned_communities

+++ src/stores/cashu/mint_mgmt.rs
@@ -1092,7 +1092,7 @@ pub async fn consolidate_proofs(
             }
             #[cfg(not(feature = "web"))]
             {
-                crate::platform::timer::sleep_ms(delay_ms as u64).await;
+                crate::platform::timer::sleep_ms(delay_ms).await;
             }
             log::info!("Retrying token event publish (attempt {})", attempt + 1);
         }
